requirements.txt
antspyx==0.4.2
catboost==1.2.5
dask==2024.5.2
imgaug==0.4.0
hdbscan==0.8.36
imblearn==0.0
imzy==0.2.6
lightgbm==4.3.0
matplotlib==3.8.4
matplotlib_venn==0.11.10
numba==0.59.1
numpy==1.26.4
opencv-contrib-python==4.10.0.82
pandas==2.1.4
Pillow==10.3.0
pyimzml==1.5.4
scikit-image==0.23.2
scikit_learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
shap==0.42.1
snakemake==8.13.0
streamlit==1.35.0
tifffile==2024.5.22
tqdm==4.66.4
umap-learn==0.5.6
xgboost==2.0.3


